Refactored admin blueprint
